backend/requirements.txt
# Githaf Chatbot Backend - Production Dependencies
# Use loose version pinning to avoid dependency conflicts
# For exact reproducible builds, use: pip freeze > requirements-lock.txt

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi>=0.110.0,<0.120.0
uvicorn[standard]>=0.27.0,<0.30.0
python-multipart>=0.0.9,<0.1.0
python-dotenv>=1.0.0,<2.0.0

# ============================================================================
# DATA VALIDATION
# ============================================================================
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
email-validator>=2.3.0,<3.0.0

# ============================================================================
# DATABASE & STORAGE
# ============================================================================
supabase>=2.4.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
PyJWT>=2.10.0,<3.0.0
bcrypt>=3.2.0,<4.0.0  # Pin to 3.x for passlib 1.7.4 compatibility
cryptography>=46.0.0,<47.0.0

# ============================================================================
# AI/ML - LLM & EMBEDDINGS
# ============================================================================
groq>=0.4.0,<0.5.0
sentence-transformers>=2.5.0,<3.0.0
langchain>=0.1.11,<0.2.0
langchain-community>=0.0.25,<0.1.0
langchain-core>=0.1.53,<0.2.0
langchain-text-splitters>=0.0.2,<0.1.0

# ============================================================================
# AI/ML - CORE LIBRARIES
# ============================================================================
torch>=2.2.0,<3.0.0
transformers>=4.57.0,<5.0.0
tokenizers>=0.22.0,<0.23.0
huggingface-hub>=0.35.0,<0.36.0
safetensors>=0.6.0,<0.7.0

# ============================================================================
# SCIENTIFIC COMPUTING
# ============================================================================
numpy>=1.26.0,<2.0.0
scipy>=1.10.0,<2.0.0  # 1.16+ requires Python 3.11+
scikit-learn>=1.5.0,<2.0.0  # 1.7+ requires Python 3.11+
joblib>=1.5.0,<2.0.0

# ============================================================================
# FILE PROCESSING
# ============================================================================
pypdf>=4.0.0,<5.0.0
python-docx>=1.1.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.1.0,<6.0.0
playwright>=1.40.0,<2.0.0  # Web scraping with browser automation

# ============================================================================
# HTTP CLIENTS & NETWORKING
# ============================================================================
httpx>=0.25.0,<0.28.0  # Compatible with supabase 2.x
aiohttp>=3.12.0,<4.0.0
requests>=2.31.0,<3.0.0

# ============================================================================
# SCHEDULING & BACKGROUND TASKS
# ============================================================================
APScheduler>=3.10.0,<4.0.0

# ============================================================================
# RATE LIMITING & MIDDLEWARE
# ============================================================================
slowapi>=0.1.9,<0.2.0
limits>=5.6.0,<6.0.0

# ============================================================================
# UTILITIES
# ============================================================================
tenacity>=8.5.0,<9.0.0
python-dateutil>=2.9.0,<3.0.0
pytz>=2024.1
tzdata>=2024.1

# ============================================================================
# PRODUCTION SERVER (Unix/Linux only)
# ============================================================================
gunicorn>=21.2.0,<22.0.0; sys_platform != 'win32'

# ============================================================================
# GOOGLE CLOUD DEVELOPMENT TOOLS
# ============================================================================
google-api-python-client
google-auth-oauthlib

# ============================================================================
# DEPENDENCIES (automatically resolved by pip)
# ============================================================================
# The following are dependencies of packages above.
# They're listed here to ensure compatibility, but pip will resolve them.
# You can remove these if you trust pip's dependency resolution.

typing-extensions>=4.12.0  # Required by multiple packages - using max compatible version
anyio>=4.11.0
aiosignal>=1.4.0
async-timeout>=4.0.0; python_version < "3.11"  # asyncio-timeout doesn't exist, use async-timeout
attrs>=25.0.0
certifi>=2024.8.0
charset-normalizer>=3.4.0
click>=8.3.0
colorama>=0.4.6
dnspython>=2.8.0
filelock>=3.19.0
frozenlist>=1.7.0
fsspec>=2024.9.0
greenlet>=3.2.0
h11>=0.16.0
httpcore>=1.0.0
idna>=3.10
iniconfig>=2.1.0
Jinja2>=3.1.0
jsonpatch>=1.33
jsonpointer>=3.0.0
MarkupSafe>=3.0.0
multidict>=6.6.0
packaging>=23.2
pillow>=11.3.0
regex>=2024.9.0
six>=1.17.0
sniffio>=1.3.0
soupsieve>=2.8
tqdm>=4.67.0
urllib3>=2.4.0
websockets>=15.0.0  # Required for realtime/supabase async client
wrapt>=1.17.0
yarl>=1.20.0
